+++ python/cell_reader.py
@@ -7,6 +7,7 @@ import pickle
 import numpy as np
 
 from loom_reader import LoomReader
+from transform import apply_transform, get_transform
 
 
 def octagon(poly):
@@ -71,29 +72,6 @@ def mean_coord(coords):
 
     '''
     return np.mean(coords, axis=0).tolist()
-
-
-def get_transform(metadata):
-    xy_array = np.array([cell['xy'] for cell in metadata.values()])
-    max_xy = np.max(xy_array, axis=0)
-    min_xy = np.min(xy_array, axis=0)
-    domain = 2000
-    # Limited by raster-based picking in deck.gl:
-    # https://github.com/uber/deck.gl/issues/2658#issuecomment-463293063
-    return {
-        'x_shift': - (max_xy[0] + min_xy[0]) / 2,
-        'y_shift': - (max_xy[1] + min_xy[1]) / 2,
-        'x_scale': domain / (max_xy[0] - min_xy[0]),
-        'y_scale': domain / (max_xy[1] - min_xy[1])
-    }
-
-
-def apply_transform(transform, xy):
-    # TODO: Rounding should be based on precision of input.
-    return [
-        round((xy[0] + transform['x_shift']) * transform['x_scale'], 2),
-        round((xy[1] + transform['y_shift']) * transform['y_scale'], 2)
-    ]
 
 
 if __name__ == '__main__':
